automation/requirements.txt
aiosmtplib==2.0.1
APScheduler==3.9.1.post1
async-generator==1.10
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==2.0.12
dnspython==2.3.0
etext==0.0.1
h11==0.14.0
idna==3.4
outcome==1.2.0
packaging==23.0
Pillow==9.4.0
PyJWT==2.6.0
pymongo==4.3.3
PySocks==1.7.1
python-dotenv==0.21.0
pytz==2022.7
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
selenium==4.7.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
twilio==7.16.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.13
webdriver-manager==3.8.5
wsproto==1.2.0
